net: lib: lwm2m_client_utils: common set of objects
A set of LwM2M objects found from most of the client applications

* Firmware object
* Connection monitoring object
* Device object
* Location object
* Security object

Adds tutorial which tells how to take the library into use and how to
create new objects and resources
