+++ doc/nrf/nrf.doxyfile.in
@@ -1981,6 +1981,11 @@ PREDEFINED             = "CONFIG_SYS_CLOCK_EXISTS=y" \
                          "CONFIG_PROFILER=y" \
 			 "CONFIG_NFC_TNEP_TAG=y" \
 			 "EVENT_TYPE_DECLARE(x)=" \
+			 "CONFIG_LWM2M_CLIENT_UTILS_LOCATION_OBJ_SUPPORT=y" \
+			 "CONFIG_LWM2M_CLIENT_UTILS_FIRMWARE_UPDATE_OBJ_SUPPORT=y" \
+			 "CONFIG_LWM2M_CLIENT_UTILS_CONN_MON_OBJ_SUPPORT=y" \
+			 "CONFIG_LWM2M_CLIENT_UTILS_DEVICE_OBJ_SUPPORT=y" \
+			 "CONFIG_LWM2M_CLIENT_UTILS_SECURITY_OBJ_SUPPORT=y"
 
 # If the MACRO_EXPANSION and EXPAND_ONLY_PREDEF tags are set to YES then this
 # tag can be used to specify a list of macro names that should be expanded. The
