+++ samples/nrf9160/lwm2m_client/CMakeLists.txt
@@ -12,7 +12,6 @@ zephyr_compile_definitions(PROJECT_NAME=${PROJECT_NAME})
 
 # NORDIC SDK APP START
 target_sources(app PRIVATE src/main.c)
-target_sources(app PRIVATE src/settings.c)
 zephyr_include_directories(src)
 # NORDIC SDK APP END
