+++ samples/nrf9160/lwm2m_client/README.rst
@@ -9,6 +9,7 @@ nRF9160: LwM2M Client
 
 The LwM2M Client demonstrates how to use Lightweight Machine to Machine (`LwM2M`_) protocol to connect an nRF9160 DK to an LwM2M server such as `Leshan Demo Server`_ via LTE.
 Once connected, the device can be queried to obtain GPS and sensor data, and to retrieve information about the modem.
+This sample uses the :ref:`lib_lwm2m_client_utils` library.
 
 Requirements
 ************
@@ -90,6 +91,15 @@ Check and configure the following configuration options for the sample:
    The sample configuration specifies the LWM2M Server to be used.
    In this sample, you can set this option to ``leshan.eclipseprojects.io`` (`public Leshan Demo Server`_).
 
+
+Additional configurations
+=========================
+
+Check and configure the following library options that are used by the sample:
+
+* :option:`CONFIG_LWM2M_CLIENT_UTILS` - Enables the utility library :ref:`lib_lwm2m_client_utils`, which initializes a predefined set of objects.
+
+* :option:`CONFIG_LWM2M_IPSO_SUPPORT` - Enables support for different IPSO objects. If you are disabling this option, you must disable all the individual configuration options that enables the sensor.
 
 
 Configuration files
