+++ samples/nrf9160/lwm2m_client/prj.conf
@@ -12,15 +12,13 @@ CONFIG_NET_SOCKETS=y
 CONFIG_NET_SOCKETS_OFFLOAD=y
 
 # LwM2M
+CONFIG_LWM2M_CLIENT_UTILS=y
+CONFIG_LWM2M_CLIENT_UTILS_LOG_LEVEL_DBG=y
 CONFIG_LWM2M=y
 CONFIG_LWM2M_DNS_SUPPORT=y
 CONFIG_LWM2M_RW_JSON_SUPPORT=n
 CONFIG_LWM2M_SERVER_DEFAULT_PMIN=1
 CONFIG_LWM2M_SERVER_DEFAULT_PMAX=300
-CONFIG_LWM2M_CONN_MON_OBJ_SUPPORT=y
-CONFIG_LWM2M_CONN_MON_BEARER_MAX=2
-CONFIG_LWM2M_FIRMWARE_UPDATE_OBJ_SUPPORT=y
-CONFIG_LWM2M_LOCATION_OBJ_SUPPORT=y
 CONFIG_LWM2M_IPSO_SUPPORT=y
 CONFIG_LWM2M_IPSO_TEMP_SENSOR=y
 CONFIG_LWM2M_IPSO_LIGHT_CONTROL=y
@@ -59,15 +57,6 @@ CONFIG_COAP_EXTENDED_OPTIONS_LEN_VALUE=40
 # UI settings
 CONFIG_UI_BUTTON=y
 CONFIG_UI_BUZZER=y
-
-# FOTA
-CONFIG_DFU_TARGET=y
-CONFIG_BOOTLOADER_MCUBOOT=y
-CONFIG_IMG_MANAGER=y
-CONFIG_MCUBOOT_IMG_MANAGER=y
-CONFIG_IMG_ERASE_PROGRESSIVELY=y
-CONFIG_FLASH=y
-CONFIG_MPU_ALLOW_FLASH_WRITE=y
 
 # Enable temp sensor
 CONFIG_SENSOR=y
