+++ samples/nrf9160/lwm2m_client/src/config.h
@@ -5,8 +5,8 @@
  */
 
 #if defined(CONFIG_LWM2M_DTLS_SUPPORT)
-static char client_psk[] = "000102030405060708090a0b0c0d0e0f";
+static char client_psk[] = CONFIG_APP_LWM2M_PSK;
 
-#define SERVER_TLS_TAG 35724861
-#define BOOTSTRAP_TLS_TAG 35724862
+#define SERVER_TLS_TAG CONFIG_APP_SERVER_TLS_TAG
+#define BOOTSTRAP_TLS_TAG CONFIG_APP_BOOTSTRAP_TLS_TAG
 #endif /* defined(CONFIG_LWM2M_DTLS_SUPPORT) */
