+++ samples/nrf9160/lwm2m_client/src/lwm2m/CMakeLists.txt
@@ -6,17 +6,6 @@
 
 zephyr_include_directories(include)
 target_sources(app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/lwm2m_device.c)
-target_sources(app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/lwm2m_security.c)
-
-target_sources_ifdef(CONFIG_LWM2M_LOCATION_OBJ_SUPPORT
-	app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/lwm2m_location.c
-	)
-target_sources_ifdef(CONFIG_LWM2M_FIRMWARE_UPDATE_OBJ_SUPPORT
-	app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/lwm2m_firmware.c
-	)
-target_sources_ifdef(CONFIG_LWM2M_CONN_MON_OBJ_SUPPORT
-	app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/lwm2m_connmon.c
-	)
 target_sources_ifdef(CONFIG_LWM2M_IPSO_LIGHT_CONTROL
 	app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/lwm2m_light_control.c
 	)
