+++ samples/nrf9160/lwm2m_client/src/lwm2m/include/lwm2m_client_app.h
@@ -4,8 +4,8 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
  */
 
-#ifndef LWM2M_CLIENT_H__
-#define LWM2M_CLIENT_H__
+#ifndef LWM2M_CLIENT_APP_H__
+#define LWM2M_CLIENT_APP_H__
 
 #include <zephyr.h>
 #include <net/lwm2m.h>
@@ -16,26 +16,8 @@
 extern "C" {
 #endif
 
-/* Security */
-int lwm2m_init_security(struct lwm2m_ctx *ctx, char *endpoint);
-
 /* Device */
-int lwm2m_init_device(char *serial_num);
-
-/* Location */
-int lwm2m_init_location(void);
-
-#if defined(CONFIG_LWM2M_FIRMWARE_UPDATE_OBJ_SUPPORT)
-void *firmware_read_cb(uint16_t obj_inst_id, size_t *data_len);
-int lwm2m_init_firmware(void);
-int lwm2m_init_image(void);
-int lwm2m_verify_modem_fw_update(void);
-#endif
-
-#if defined(CONFIG_LWM2M_CONN_MON_OBJ_SUPPORT)
-int lwm2m_init_connmon(void);
-int lwm2m_update_connmon(void);
-#endif
+int lwm2m_app_init_device(char *serial_num);
 
 #if defined(CONFIG_LWM2M_IPSO_LIGHT_CONTROL)
 int lwm2m_init_light_control(void);
@@ -65,4 +47,4 @@ int lwm2m_init_accel(void);
 }
 #endif
 
-#endif /* LWM2M_CLIENT_H__ */
+#endif /* LWM2M_CLIENT_APP_H__ */
