+++ samples/nrf9160/lwm2m_client/src/lwm2m/lwm2m_buzzer.c
@@ -7,6 +7,7 @@
 #include <zephyr.h>
 #include <net/lwm2m.h>
 
+#include "lwm2m_client_app.h"
 #include "ui.h"
 
 #include <logging/log.h>
