+++ subsys/net/lib/CMakeLists.txt
@@ -17,3 +17,4 @@ add_subdirectory_ifdef(CONFIG_ZZHC zzhc)
 add_subdirectory_ifdef(CONFIG_ICAL_PARSER icalendar_parser)
 add_subdirectory_ifdef(CONFIG_FTP_CLIENT ftp_client)
 add_subdirectory_ifdef(CONFIG_COAP_UTILS coap_utils)
+add_subdirectory_ifdef(CONFIG_LWM2M_CLIENT_UTILS lwm2m_client_utils)
