+++ subsys/net/lib/Kconfig
@@ -18,5 +18,6 @@ rsource "zzhc/Kconfig"
 rsource "icalendar_parser/Kconfig"
 rsource "ftp_client/Kconfig"
 rsource "coap_utils/Kconfig"
+rsource "lwm2m_client_utils/Kconfig"
 
 endmenu
