+++ subsys/net/lib/lwm2m_client_utils/lwm2m/lwm2m_firmware.c
@@ -16,17 +16,16 @@
 #include <modem/nrf_modem_lib.h>
 #include <power/reboot.h>
 
-#include "settings.h"
+#include <net/lwm2m_client_utils.h>
+#include <net/lwm2m_client_utils_fota.h>
 
 #include <logging/log.h>
-LOG_MODULE_REGISTER(app_lwm2m_firmware, CONFIG_APP_LOG_LEVEL);
+LOG_MODULE_REGISTER(lwm2m_firmware, CONFIG_LWM2M_CLIENT_UTILS_LOG_LEVEL);
 
 #define BYTE_PROGRESS_STEP (1024 * 10)
 #define REBOOT_DELAY       K_SECONDS(1)
 
-#if defined(CONFIG_LWM2M_FIRMWARE_UPDATE_OBJ_SUPPORT)
 static uint8_t firmware_buf[CONFIG_LWM2M_COAP_BLOCK_SIZE];
-#endif
 
 #ifdef CONFIG_DFU_TARGET_MCUBOOT
 static uint8_t mcuboot_buf[CONFIG_APP_MCUBOOT_FLASH_BUF_SZ];
@@ -44,7 +43,6 @@ static void reboot_work_handler(struct k_work *work)
 	sys_reboot(SYS_REBOOT_COLD);
 }
 
-#if defined(CONFIG_LWM2M_FIRMWARE_UPDATE_OBJ_SUPPORT)
 static int firmware_update_cb(uint16_t obj_inst_id, uint8_t *args,
 			    uint16_t args_len)
 {
@@ -206,18 +204,15 @@ cleanup:
 
 	return ret;
 }
-#endif
 
 int lwm2m_init_firmware(void)
 {
 	k_work_init_delayable(&reboot_work, reboot_work_handler);
 
-#if defined(CONFIG_LWM2M_FIRMWARE_UPDATE_OBJ_SUPPORT)
 	lwm2m_firmware_set_update_cb(firmware_update_cb);
 	/* setup data buffer for block-wise transfer */
 	lwm2m_engine_register_pre_write_callback("5/0/0", firmware_get_buf);
 	lwm2m_firmware_set_write_cb(firmware_block_received_cb);
-#endif
 
 	return 0;
 }
