+++ subsys/net/lib/lwm2m_client_utils/lwm2m/lwm2m_security.c
@@ -6,9 +6,10 @@
 
 #include <zephyr.h>
 #include <net/lwm2m.h>
+#include <net/lwm2m_client_utils.h>
 
 #include <logging/log.h>
-LOG_MODULE_REGISTER(app_lwm2m_security, CONFIG_APP_LOG_LEVEL);
+LOG_MODULE_REGISTER(lwm2m_security, CONFIG_LWM2M_CLIENT_UTILS_LOG_LEVEL);
 
 #include "config.h"
